Create registerpollingmegashear
This is a work in process, a part of an embedded system design for a metal shearing station. This section of code runs a subroutine, (state machine within a state machine) which engage some pneumatic 'pushers' that engage when a chain drive feeder can no longer get the last 12" of sheet steel up against the back gauge of a steel shear. This subroutine uses proximity sensors to control pneumatic pushers.