Create FullCalibrationSketchHX711LoadCell
This is functional and used to calibrate a load cell to the HX711. It is used to calibrate the scale through a serial input, but will likely be replaced with with a mechanical taring function... or leave tare weight equals -0.03 to ensure that 50.03 is over weight by 6/100ths of a pound.